DWIM when typing RET on a column that references a foreign table.
A column value that references a foreign table will be shown in blue
using the default faces. Typing RET on the value jumps to the referenced
row in the referenced table, instead of prompting to edit the value. You
can edit nonetheless by typing w to open the widget-based editing buffer.

This feature following a suggestion by @mbork
https://mbork.pl/2024-06-24_PGmacs_-_editing_PostgreSQL_databases_from_Emacs