New features: WHERE filtering and multirow delete.
WHERE filtering makes it possible to only display rows in a PostgreSQL table that match an SQL WHERE
clause.

Multirow delete is inspired by dired functionality for marking files for deletion. Press d to mark a
row for deletion; it will be colored in a (currently somewhat aggressive) red background color.
Press u to unmark a row. Press x to delete (expunge) the marked rows.

Multirow delete functionality has been tested both for tables with a simple primary key, and for
tables with a multicolumn primary key. It leads to some edge cases in interactions with the
insertion of new rows (including copy-paste of rows) which are currently insufficiently tested:
interspersing multirow deletion and insertion commands is not currently recommended.