Functions targeted for interactive use ignore their arguments.
Instead of accepting the current row as an argument, they all
use pgmacstbl-current-object. This will allow them to be used
as a button action, or using the :action constructor argument
for a pgmacstbl, or for a :keymap constructor argument (which
we will switch to to allow keybindings to be customized by
the user).